server files added for dev and prod
